Fix special handling of usage message
Normally the add-on will be invoked as a subcommand (like
`todo.sh mit foo`), meaning that the actual mit-action, along with its
various arguments, will be placed in ARGV[1] and higher.

The only exception to this case is when the add-on is invoked through
`todo.sh help`. This command will iterate through all available add-ons
and invoke them like `mit usage`, meaning that the action will be
present in ARGV[0].