Refactor to make procedure a bit more explicit
While the code is compact, in the context of a script like this, I think
there's value in "surfacing" implementation details like what output
stream and what exit code is used. Of course, this is a case-by-case
thing, but here it feels right make things a bit more explicit.

Additionally, by doing the extraction in this patch, the private methods
that show messages have been reduced to just generating messages, making
them nicely focused.