Add basic printing of MITs
The main purpose of this patch is to enable printing of MITs when the
add-on is invoked through `todo.sh mit`.

Here's a summary of the changes introduced in this patch:

* Add basic parsing of the TODO_FILE, filtering out MITs.
* Add a skeleton class for representing MITs.
* Amend the Executable helper to enable tweaking ENV with provided
  hashes. This allows us to inject the location of an alternate
  "todo.txt" file, by placing an alternate path in the `$TODO_FILE`
  ENV-variable. Ensure that the ENV is restored to its original state
  after the Executable is finished running its command.
* Add tests for verifying the presence of the MITs and the absence of
  the non-MITs.
* Update the README to include information about the gem requirements
  for running the tests. While "rspec" was already required (and this
  should have been in the README), this patch extends the requirements
  to include "timecop".

I've opted to not check the formatting of the output too closely; while
the intention is to replicate the original add-on for the most part,
I don't yet feel like rigidly enforcing the style through tests.