Refactor test helpers for test readability
The two helper methods in the `Listing MITs` tests have plenty of
duplicated code, but more importantly the part which make them differ
is the most important part have immediately readable up in the actual
test code: the TODOs.

Refactor the helper methods to take the TODO lines as an argument and
move the TODO lines to the individual tests.