Replace Timecop with ENV variable hack
Since all the testing is done with the Executable helper, and that
helper in turn makes use of `popen3`, Timecop will not work for the
purpose of freezing the time to a specific date. This is due to the fact
that `popen3` runs the script in a child process, separate from the
environment in which Timecop has applied its changes.

To be able to inject a specific date for testing, read the `FIXED_DATE`
environment variable and parse it when starting the script.

Remove Timecop dependency (adjust README correspondingly).