Change helper method to require date for clarity
Currently one has to scroll down to the helper method definition to
figure out what the "fixed time" is, meaning that the TODOs defined
above the usage of the helper (which are passed in as an argument) make
much less sense, since they lack the context of what date they're
checked against. Apart from that, it's nice to have the flexibility of
setting a different date.

Adjust the helper method to take a date string as an argument.