Add single file script generator
Currently the executable placed in the `bin/` directory does little more
than require the rest of the source code placed in the `lib/` directory.

Since most of the add-ons for `todo.sh` are packaged as single file
scripts, this patch adds a generator that combines all of the source
files into one script.

Update the README with instructions for how to make use of this
generator, along with updated installation instructions.