todo.txt-mit 0.1.2
This release primarily deals with the separation of the executable into
separate files, for the purpose of simplifying testing and development.

This has the side effect that one can no longer just copy the executable
`bin/mit` in the add-on folder. A new installation procedure has been
detailed in the README.