Enable addition of MITs with "today"
Apart from adding the ability to add MITs with the string "today" as the
date, a minor issue was fixed in `TodoFileMutator`:

Currently the MIT that is written to the `TODO_FILE` uses the date as
passed from the user, not using the parsed date. This is fine for when
one uses the YYYY.MM.DD dates, but breaks for the new date format
introduced in this patch.

`parsed_date` is what it should have been in the first place, anyway, so
change it to that.