Inline simple extractions
Since I've already made the decision that the case-statement should be
more explicit (0bde5554), eliminate extractions for rather simple logic.

In the case of `usage_requested_from_todo_help?`, the extraction is
warranted due to the large explanatory comment associated with the
logic; in the cases touched by this patch, the logic is obvious in
nature.