Enable addition of MITs with weekday names
Add the ability to add MITs with both long or short weekday names.
Examples include "monday", "saturday" or "sat".

Refactor the `DateParser#parse` method to only describe the sequence of
parsing strategies. Extract the procedure of the various parsing
strategies into private methods. Prefix the names of these private
methods with "maybe" to imply that one can either get a result or `nil`.

I'm a bit unsure about whether `DateParser#parse` potentially returning
`nil` warrants renaming the method to `DateParser#maybe_parse`. For now,
I think I'll leave the name as it is.

This patch marks a departure from how the original add-on works:

If today is a Tuesday (for instance), then adding a MIT with the weekday
"tuesday" will add a MIT for the Tuesday in the following week, instead
of today. The behavior in the original add-on comes from shelling out to
the `date` application, which would have considered "tuesday" to be
today.

While there's probably a good reason for the `date` application behaving
as described above, for the purpose of this add-on it is unhelpful:
I will never refer to today with the weekday; conversely if I say
the weekday of today then I will always mean the day one week from now.