Refactor printing all MITs
In anticipation of the addition of filtering of MITs by context, move
all of the logic related to composing a MIT listing into one class.

MITListFormatter already does most of the work related to the task of
listing MITs, so only a few minor adjustments are necessary to make it
ready context filtering.