Change interface for adding MITs
While fully working and tested, the interface for adding MITs (the `add`
action) is inconsistent with how the original add-on works: the original
add-on does not require the `add` action to be written.

While one could argue that it would be more consistent if one had action
names included in all the commands (which, to be fair, would necessitate
tweaking context filtering as well), for the first major release this
project should mostly behave like the original add-on.