Add ability to use relative dates
When adding or changing the date of MITs, one can currently only use
absolute dates (in the format "YYYY.MM.DD") or weekday names (along with
the words "today" and "tomorrow").

This patch enables the use of relative dates such as "1d", "2w" and "3m"
(one day, two weeks and three months, respectively).