Format output dates when using `mv`
When moving a MIT from one date to another, a message is printed along
the lines of "TODO: 'foobar' moved to YYYY-MM-DD".

There's already a `DateFormatter` present, but it does not handle dates
in the past. This limitation was put in place as a redundant safeguard:
the `MITListPrinter` bundles together all the MITs in the past under the
category of "Past due", which means that `DateFormatter` should not have
been called to format dates in the past. It was redundant because the
filtering made by `MITListPrinter` ensured that this safeguard was never
called.

To make `DateFormatter` work as a formatter for the `mv` operation,
change the following things:

* Make the addition of a colon at the end of the formatted date
  configurable.
* Make the capitalization of "today" and "tomorrow" configurable.
  Calling String#capitalize over in `MITListPrinter` on the formatted
  date is not an option since that method would downcase weekday names
  if they're not at the start of the formatted date.
* Make `DateFormatter` format dates in the past with by just calling
  Date#to_s.

The changes introduced in this patch to the implementation necessitate
some changes to the tests:

The tests currently expect that usage of the `mv` action result in
a message being printed to STDOUT that contains the new date of the MIT
in "YYYY-MM-DD"-form. Since `DateFormatter` does not always include the
date in said form (compare "today", "Tuesday" and "Thursday, 2 weeks
from now, 2017-02-16"), the tests need to be adjusted to either expect
words such as "tomorrow" or have more lenient expectations.