Create variable with explanatory name
The line describing what should be printed upon creation of a MIT should
print the ID of the new TODO with the word "added" following it. The
fact that TODOs are added at the end of the TODO file by the `todo.sh`
application and that subsequently the ID is the same as the number of
lines in the TODO_FILE is something that isn't immediately clear.

Add an intermediary variable with an appropriate name to make things
a bit more clear.