Make "type transitions" more prominent
Currently the objects that are passed in as parameters to the public
methods of `TodoFileMutator` instances (specifically the `task_id` and
the `date`) will always be instances of String. Both `task_id` and
`date` will need to go through layers of checks before they can be
turned into more useful objects.

The fact that `task_id` and `date` are strings should be made more
clear. These string objects are either passed into a parser or
a validation method and then the return values are mainly used; I think
the intention of not using the parameter objects in their string form
can be made more explicit if the "_string" suffix is appended to both of
their names.

Additionally, the private method `TodoFileMutator#maybe_find_task` both
checks the validity of the `task_id` string and returns a task if the
string was valid. While this resembles the interaction with
`DateParser#parse`, it should be changed:

The task is obtained by calling the `maybe_find_task` method, which
implicitly means that the `task_id` string has been validated and can
safely be coerced with String#to_i further down. This validation should
not be implicit, considering the task ID is as important as the task
itself.