todo.txt-mit 1.0.0
This release signifies the point where I consider the add-on to be more
or less complete.

The major version bump is also caused by a breaking API change:

The addition of new MITs was previously done with the `add` sub-command.
The original add-on didn't use a sub-command for adding MITs. While
I see some value in having a dedicated sub-command, I've opted to
mimic the interface used by the original add-on, for compatibility's
sake.

The major addition in this release is the ability to use relative dates
when moving or adding MITs. For instance, if one would like to add a MIT
three days into the future one could invoke the add-on with the
following command:

  todo.sh mit 3d "foo @bar +baz"

If one wanted to move a MIT two weeks into the future, this command
would be used:

  todo.sh mit mv 123 2w

The relative dates are expressed as an integer, followed by either "d"
for days, "w" for weeks, or "m" for months.

Finally, a number of refactorings have been included to make the code
a bit more "1.0-ready".