Print error if date argument is missing
Currently if the command `todo.sh mit mv 123 tomorrow` has the date
argument missing ("tomorrow", in this case), the application crashes,
since the code expects to work with a string, while the missing argument
results in a nil.

Check if the argument is missing and print an informative error if that
is the case.