Make liberal date input a documented feature
In 59e38643 the `Date.parse` method was used as the only strategy for
parsing date strings. The usage of this method was probably meant to be
temporary, serving as a stand-in for the (at the time) unimplemented
parsing strategies (such as the MIT date format, "YYYY.MM.DD"). Instead,
the usage of `Date.parse` persisted, despite there being no tests
requiring its functionality.

Recently, however, I discovered the usefulness of having the very
liberal parsing of `Date.parse` available as an alternative to the
relative dates, allowing expressions such as "23rd" or "25 mar" to
be used as date strings. Subsequently:

Document the liberal parsing as a feature in the usage message and add
tests that verify that the functionality is present.