Add MIT copying feature
This patch adds the `cp` sub-command that takes a task and adds it once
more on the specified date as a MIT.

When it comes to the implementation, the task specified by ID is simply
cleaned up and then passed along to the already existing
TodoFileMutator#add_mit method.

The tests added are largely a copy of the tests present in
`spec/adding_mits_spec.rb`, with some tweaks made for verifying that the
copying is done properly. The duplication is notable and could be
avoided if some of the functionality was verified with unit tests (most
notably the date string handling). With this said, the tests do verify
the correct working of the `cp` sub-command.

Update the usage message to include information about `cp`.