Chomp task to ensure one line output
If the task is not chomped at some point prior to sending it off to
TodoFileMutator#add_mit, then outputting the task content will result in
an extra newline.

I don't think that TodoFileMutator#add_mit should care for tasks that
for some reason have trailing newlines or are multi-line (even weirder
and should not happen). Therefore I suspect that chomping the task in
TodoFileMutator#copy_mit is a good place to do it.